+++ ethers-core/src/jvmSharedMain/kotlin/io/ethers/core/types/transaction/TxDynamicFee.kt
@@ -1,5 +1,7 @@
 package io.ethers.core.types.transaction
 
+import io.ethers.bigint.BigInt
+import io.ethers.bigint.BigInts
 import io.ethers.core.types.AccessList
 import io.ethers.core.types.Address
 import io.ethers.core.types.Authorization
@@ -9,7 +11,6 @@ import io.ethers.core.types.Signature
 import io.ethers.rlp.RlpDecodable
 import io.ethers.rlp.RlpDecoder
 import io.ethers.rlp.RlpEncoder
-import java.math.BigInteger
 
 /**
  * EIP-1559 transaction with optional access list and [gasFeeCap]/[gasTipCap] instead of [gasPrice].
@@ -18,11 +19,11 @@ import java.math.BigInteger
  * */
 data class TxDynamicFee(
     override val to: Address?,
-    override val value: BigInteger,
+    override val value: BigInt,
     override val nonce: Long,
     override val gas: Long,
-    override val gasFeeCap: BigInteger,
-    override val gasTipCap: BigInteger,
+    override val gasFeeCap: BigInt,
+    override val gasTipCap: BigInt,
     override val data: Bytes?,
     override val chainId: Long,
     override val accessList: List<AccessList.Item>,
@@ -33,13 +34,13 @@ data class TxDynamicFee(
         }
     }
 
-    override val gasPrice: BigInteger
+    override val gasPrice: BigInt
         get() = gasFeeCap
 
     override val type: TxType
         get() = TxType.DynamicFee
 
-    override val blobFeeCap: BigInteger?
+    override val blobFeeCap: BigInt?
         get() = null
 
     override val blobVersionedHashes: List<Hash>?
